requirements.txt
annotated-types==0.6.0
anyio==4.0.0
asgiref==3.7.2
blinker==1.6.3
certifi==2023.7.22
charset-normalizer==3.3.1
click==8.1.7
deprecation==2.1.0
Django==4.2.6
django-admin==2.0.2
django-excel-response2==3.0.6
django-six==1.0.5
excel-base==1.0.4
exceptiongroup==1.1.3
Flask==3.0.0
gotrue==1.2.0
gunicorn==21.2.0
h11==0.14.0
httpcore==0.17.3
httpx==0.24.1
idna==3.4
isoweek==1.3.3
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
packaging==23.2
postgrest==0.11.0
pydantic==2.4.2
pydantic_core==2.10.1
python-dateutil==2.8.2
python-decouple==3.8
python-dotenv==1.0.0
realtime==1.0.0
requests==2.31.0
screen==1.0.1
six==1.16.0
sniffio==1.3.0
sqlparse==0.4.4
storage3==0.6.1
StrEnum==0.4.15
supabase==1.2.0
supafunc==0.2.3
TimeConvert==3.0.13
typing_extensions==4.8.0
tzlocal==5.2
urllib3==2.0.7
websockets==10.4
Werkzeug==3.0.1
xlwt==1.3.0
